Dark theme/OLED-compatible version of floor.png
May even be suitable for the Light theme of the skin. No downsides to this detected in Dark theme, as the white version of floor.png seems wholly unsuited for the Dark theme and OLED - though it may slightly intensify the darkening effect of the overlay - which can be compensated by the 'Brightness' (opacity) settings in the skin.